+++ backend/sokdak/src/test/java/com/wooteco/sokdak/auth/controller/AuthControllerTest.java
@@ -87,23 +87,4 @@ class AuthControllerTest extends ControllerTest {
                 .apply(document("refresh/success"))
                 .statusCode(HttpStatus.NO_CONTENT.value());
     }
-
-    @DisplayName("인증코드가 일치하지 않거나 만료되었으면 400 반환")
-    @Test
-    void verifyAuthCode_Exception_different() {
-        VerificationRequest verificationRequest = new VerificationRequest("test@gmail.com", "a1b2c3");
-        doThrow(new InvalidAuthCodeException())
-                .when(authService)
-                .verifyAuthCode(any());
-
-        restDocs
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(verificationRequest)
-                .when().post("/members/signup/email/verification")
-                .then().log().all()
-                .assertThat()
-                .body("message", equalTo("잘못된 인증번호입니다."))
-                .apply(document("member/verification/fail"))
-                .statusCode(HttpStatus.BAD_REQUEST.value());
-    }
 }
