+++ backend/sokdak/src/test/java/com/wooteco/sokdak/support/token/JwtTokenProviderTest.java
@@ -1,12 +1,9 @@
 package com.wooteco.sokdak.support.token;
 
 import static com.wooteco.sokdak.member.domain.RoleType.USER;
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.wooteco.sokdak.auth.dto.AuthInfo;
-import com.wooteco.sokdak.member.domain.RoleType;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
